tts_service/requirements.txt
# Kokoro TTS Requirements
flask
requests
pygame
pyyaml
torch
torchaudio
soundfile
numpy
